admincenter 登录功能完成

+++ admincenter-dal/src/main/java/com/mall/admincenter/dal/dao/BaseDAO.java
@@ -26,6 +26,6 @@ public interface BaseDAO<T> {
      * @param id
      * @return
      */
-    T getById(Long id);
+    T getById(Integer id);
 
 }
